feat: add SearchArticles page design
